requirements.txt
alembic==1.9.2
aniso8601==9.0.1
APIFlask==1.1.3
apispec==5.2.2
attrs==22.2.0
autopep8==1.7.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.11.0
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==4.0.1
Flask-REST==1.3
flask-restx==1.0.3
Flask-Share==0.1.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
PyJWT==2.6.0
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.3
pytest==7.2.1
python-decouple==3.7
python-dotenv==0.21.0
pytz==2022.7
qrcode==7.4.2
requests==2.28.2
share==1.0.4
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.41
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
webargs==8.2.0
Werkzeug==2.2.2